Unbreak libraries using Qt, together with moccable classes in .cpp files.
The generator that handles moccable cpp files has to actualize created
targets right away. However, it did that before calling
virtual-target.register, so, if the generator was invoked two times on
the same target, it will produce two identical targets and call .actualize
on it. The second invocation would trigger "duplicate target" error.