Add two new classes: pistream and postream.  These extend the standard istream and ostream classes to encapsulate the objects required to communicate with a system pipe (pipe and systembuf objects).  Simplifies implementation of basic_child a lot because objects are kept where they should be (they are not destroyed earlier).
Anyway, the real reason behind this change is to be able to let the user
close process descriptors at will through the close methods in these
streams.  This is important to let him, e.g., tell a process that it
shall not expect any more data through stdin.

Fixes the child regression test that did not work yesterday.