Launcher and child rototill: make these two classes only handle the three standard data streams so that they make sense under both Unix and Win32 systems (the latter cannot manage more communication flows).  The current implementation can certainly benefit from some clean up if the pipe class is rethought.
While doing this, move the existing launcher class to posix_launcher,
which ceases to work properly at the moment.  This needs to be rethought
at a later stage but I want to get the basic launcher working on the two
platforms first.

Also drop the standard_launcher class; the new launcher is easy enough
to use to make this one useless.