Move the posix_argv and win32_cmdline command_line member functions out of this class and put them in the detail namespace.
These two functions are just used for convenience to improve cohesion and
to simplify the launcher's code.  However, a command_line class should not
be forced to provide them, as was the case now.  This change unclutters the
command_line visible interface (writing Doxygen comments is great to find
these issues! :-) and eases its replacement with a user specific class if
he wants to provide it.

There is still an open question... these two functions may come handy if
the end user decides to write his own launcher class, but as they are in
the detail namespace he should not even know they exist.  Of course he
may duplicate the code (after all the functions are small), but I wonder
if there might be a better solution.  Anyway, this is a tiny issue and I
hope this is good enough for now.